UserEdit and UserDeactivate working on Profile Page
