+++ bc-react/src/style/UserLogin.scss
@@ -13,6 +13,7 @@
 
   form {
       input {
+        text-align: center;
         font-family: 'Abel', sans-serif;
         font-size: .9em;
       }
